code/team2/PirateShooter/requirements.txt
autopep8
evdev
grpcio
grpcio-tools
protobuf
Jetson.GPIO
adafruit-circuitpython-pca9685
adafruit-circuitpython-servokit
numpy